refactor: organizar imports e realocar testes para a tasta de testes
