+++ src/aluno/aluno.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AlunoService } from './aluno.service';
-import { AlunoController } from './aluno.controller';
 import { Aluno } from '../entities/aluno/aluno.entity';
-import { Inscricao } from '../entities/inscricao/inscricao.entity';
 import { Beneficio } from '../entities/beneficio/beneficio.entity';
+import { Inscricao } from '../entities/inscricao/inscricao.entity';
+import { AlunoController } from './aluno.controller';
+import { AlunoService } from './aluno.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Aluno, Inscricao, Beneficio])],
