+++ src/app.module.ts
@@ -1,16 +1,16 @@
-import { ConfigModule } from '@nestjs/config';
 import { Module } from '@nestjs/common';
-import { AuthModule } from './auth/auth.module';
-import { AlunoModule } from './aluno/aluno.module';
+import { ConfigModule } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { typeOrmConfig } from './db/db.config';
-import { EditalModule } from './edital/edital.module';
-import { MinioClientModule } from './minio/minio-client.module';
-import { InscricaoModule } from './inscricao/inscricao.module';
-import { DocumentoModule } from './documentos/documentos.module';
-import { ValidacaoModule } from './validacao/validacao.module';
+import { AlunoModule } from './aluno/aluno.module';
+import { AuthModule } from './auth/auth.module';
 import { BeneficioModule } from './beneficio/beneficio.module';
+import { typeOrmConfig } from './db/db.config';
+import { DocumentoModule } from './documentos/documentos.module';
+import { EditalModule } from './edital/edital.module';
+import { InscricaoModule } from './inscricao/inscricao.module';
+import { MinioClientModule } from './minio/minio-client.module';
 import { StepModule } from './step/step.module';
+import { ValidacaoModule } from './validacao/validacao.module';
 
 @Module({
   imports: [
