+++ src/auth/auth.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
+import { AlunoModule } from '../aluno/aluno.module';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
-import { AlunoModule } from '../aluno/aluno.module';
 
 @Module({
   imports: [AlunoModule],
