+++ src/auth/dto/updatepassword.dto.ts
@@ -1,8 +1,8 @@
+import { ApiProperty } from '@nestjs/swagger';
 import {
   IsNotEmpty,
   Matches,
 } from 'class-validator';
-import { ApiProperty } from '@nestjs/swagger';
 
 export class UpdatePasswordDto {
   @ApiProperty({
