+++ src/beneficio/beneficio.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Aluno } from '../entities/aluno/aluno.entity';
-import { BeneficioService } from './beneficio.service';
-import { BeneficioController } from './beneficio.controller';
 import { Beneficio } from '../entities/beneficio/beneficio.entity';
+import { BeneficioController } from './beneficio.controller';
+import { BeneficioService } from './beneficio.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Aluno, Beneficio])],
