+++ src/db/db.config.ts
@@ -1,5 +1,5 @@
-import { DataSource, DataSourceOptions } from 'typeorm';
 import { config } from 'dotenv';
+import { DataSource, DataSourceOptions } from 'typeorm';
 
 config();
