+++ src/documentos/documentos.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Inscricao } from 'src/entities/inscricao/inscricao.entity';
-import { Documento } from 'src/entities/documento/documento.entity';
 import { Aluno } from 'src/entities/aluno/aluno.entity';
+import { Documento } from 'src/entities/documento/documento.entity';
+import { Inscricao } from 'src/entities/inscricao/inscricao.entity';
 import { DocumentoController } from './documentos.controller';
 import { DocumentoService } from './documentos.service';
