+++ src/edital/dto/create-etapas-edital.dto.ts
@@ -1,6 +1,6 @@
+import { ApiProperty } from '@nestjs/swagger';
 import { Type } from 'class-transformer';
 import { IsDate, IsNotEmpty, IsNumber, IsString } from 'class-validator';
-import { ApiProperty } from '@nestjs/swagger';
 
 export class CreateEtapasDto {
   @ApiProperty({ description: 'Nome da etapa' })
