+++ src/edital/dto/update-edital.dto.ts
@@ -1,7 +1,7 @@
-import { IsEnum, IsNotEmpty } from 'class-validator';
 import { ApiProperty, PartialType } from '@nestjs/swagger';
-import { CreateEditalDto } from './create-edital.dto';
+import { IsEnum, IsNotEmpty } from 'class-validator';
 import { StatusEdital } from 'src/enum/enumStatusEdital';
+import { CreateEditalDto } from './create-edital.dto';
 
 export class UpdateEditalDto extends PartialType(CreateEditalDto) {
   @ApiProperty({ description: 'Status do edital', example: StatusEdital.EM_ANDAMENTO })
