+++ src/edital/edital.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { EtapaEdital } from 'src/entities/etapaEdital/etapaEdital.entity';
+import { Edital } from '../entities/edital/edital.entity';
 import { EditalController } from './edital.controller';
 import { EditalService } from './edital.service';
-import { Edital } from '../entities/edital/edital.entity';
-import { EtapaEdital } from 'src/entities/etapaEdital/etapaEdital.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Edital, EtapaEdital])],
