+++ src/validacao/validacao.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { ValidacaoService } from './validacao.service';
-import { ValidacaoController } from './validacao.controller';
 import { Validacao } from '../entities/validacao/validacao.entity';
+import { ValidacaoController } from './validacao.controller';
+import { ValidacaoService } from './validacao.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Validacao])],
